Extract changelog entries for a given version

Extracts the section of NEWS.md corresponding to a given version.

get_changes(path, version_number, verbose = TRUE)

Arguments

path

[character] Path to the package root directory.

version_number

[character] Current version number string (e.g. "1.2.3").

verbose

[logical] Whether to print current and new remote fields (default: TRUE).

Value

A character string containing the formatted changelog for the given version.

Examples

path_rjd3workspace <- system.file("rjd3workspace", package = "releaser")

get_changes(path = path_rjd3workspace, version_number = "Unreleased")
#> Reading NEWS.md from: /home/runner/work/_temp/Library/releaser/rjd3workspace
#> Extracting changes for version: Unreleased
#> Successfully extracted 6 lines of changes.
#> [1] "## Changes\n\n### Changed\n\n* Rename `.jws_add()` into `jws_add()`\n\n\n[Unreleased]: https://github.com/rjdverse/rjd3workspace/compare/v3.5.1...HEAD"
get_changes(path = path_rjd3workspace, version_number = "3.2.4")
#> Reading NEWS.md from: /home/runner/work/_temp/Library/releaser/rjd3workspace
#> Extracting changes for version: 3.2.4
#> Successfully extracted 12 lines of changes.
#> [1] "## Changes\n\n### Fixed\n\n* fixed `set_specification()` and `set_domain_specification()`\n* Correction of imports when the workspace contains no SAP.\n* `replace` parameter of `save_workspace()` corrected in R (not in Java).\n\n### Added\n\n* new function `.jsa_jresults()`, `.jread_sap()` and `.jread_workspace()` to  extract the Java object of the results of a SAItem.\n\n\n[3.2.4]: https://github.com/rjdverse/rjd3workspace/compare/v3.2.3...v3.2.4"
get_changes(path = path_rjd3workspace, version_number = "3.5.1")
#> Reading NEWS.md from: /home/runner/work/_temp/Library/releaser/rjd3workspace
#> Extracting changes for version: 3.5.1
#> Successfully extracted 27 lines of changes.
#> [1] "## Changes\n\n### Added\n\n* For compatibility reasons, some functions have been added again (`.jmp_name()`, `.jmp_sa()`, `.jmp_sa_count()`, `.jmp_sa_name()`, `.jsap_sa_count()`, `.jsap_sa_name()`)\n* New functions have deprecated old functions (`get_metadata()`, `get_ts_metadata()`, `transfer_sa_item()`)\n\n### Deprecated\n\n* Function `.jmp_sa_count()` is deprecated. Please use function `sap_sai_count()` instead.\n* Function `.jsap_sa_count()` is deprecated. Please use function `sap_sai_count()` instead.\n* Function `.jmp_name()` is deprecated. Please use function `sap_name()` instead.\n* Function `.jmp_sa()` is deprecated. Please use function `jsap_sai()` instead.\n* Function `.jmp_sa_name()` is deprecated. Please use function `sap_sai_names()` instead.\n* Function `.jsap_sa_name()` is deprecated. Please use function `sap_sai_names()` instead.\n* Function `.jsai_ts_metadata()` is deprecated. Please use function `get_ts_metadata()` instead.\n* Function `transfer_series()` is deprecated. Please use function `transfer_sa_item()` instead.\n* Function `.jsai_metadata()` is deprecated. Please use function `get_metadata()` instead.\n\n### Changed\n\n* Simplified the modification of path in `update_path()`\n\n### Removed\n\n* `add_variable()` is removed. Please use now `add_variables()`\n\n\n[3.5.1]: https://github.com/rjdverse/rjd3workspace/compare/v3.5.0...v3.5.1"